requirements.txt
spacy>=3.7.0
transformers>=4.35.0
torch>=2.0.0
sentence-transformers>=2.2.0
pandas>=2.0.0
numpy>=1.24.0
python-heideltime>=0.1.0
huggingface_hub>=0.20.0

